Adds more comments, Roles.GetRoleValue, README

+++ permissions.go
@@ -1,12 +1,16 @@
 package tinyroles
 
 // Creates Permission with Value of 2^`base`
-func NewPermission(base uint64) Permission {
+func NewPermission(base uint) Permission {
 	return permission(1 << base)
 }
 
+// Permissions interface
+// It is sealed to ensure no calculation rules was violated when `Value()` is calculated
 type Permission interface {
 	sealed()
+
+	// Returns permissions value
 	Value() uint64
 }
